- REF(base/permission,users/permissions,auth/dependencies,auth/permissions) - Moved BaseUserPermissionService from base.permission to users.permissions.user module and updated all imports
